fix: remove sample game and organize games directory
