--- src/pages/games/sample.tsx
@@ -1,6 +1,0 @@
-import GameWrapper from "@/components/games/GameWrapper";
-import ExampleGame from "@/components/games/ExampleGame";
-
-export default function SampleGamePage() {
-  return <GameWrapper GameComponent={ExampleGame} />;
-}
